Re-organize knowhere benchmark
