+++ unittest/benchmark/CMakeLists.txt
@@ -44,9 +44,9 @@ macro(benchmark_test target file)
     install(TARGETS ${target} DESTINATION unittest)
 endmacro()
 
-benchmark_test(benchmark_faiss                  benchmark_faiss.cpp)
-benchmark_test(benchmark_knowhere_binary        benchmark_knowhere_binary.cpp)
-benchmark_test(benchmark_knowhere_binary_range  benchmark_knowhere_binary_range.cpp)
-benchmark_test(benchmark_knowhere_float         benchmark_knowhere_float.cpp)
-benchmark_test(benchmark_knowhere_float_range   benchmark_knowhere_float_range.cpp)
-benchmark_test(benchmark_knowhere_perf          benchmark_knowhere_perf.cpp)
+benchmark_test(benchmark_faiss_float            hdf5/benchmark_faiss_float.cpp)
+benchmark_test(benchmark_knowhere_binary        hdf5/benchmark_knowhere_binary.cpp)
+benchmark_test(benchmark_knowhere_binary_range  hdf5/benchmark_knowhere_binary_range.cpp)
+benchmark_test(benchmark_knowhere_float         hdf5/benchmark_knowhere_float.cpp)
+benchmark_test(benchmark_knowhere_float_range   hdf5/benchmark_knowhere_float_range.cpp)
+benchmark_test(benchmark_knowhere_perf          hdf5/benchmark_knowhere_perf.cpp)
