+++ unittest/benchmark/hdf5/benchmark_hdf5.h
@@ -12,7 +12,6 @@
 #pragma once
 
 #include <assert.h>
-#include <gtest/gtest.h>
 #include <hdf5.h>
 #include <sys/time.h>
 
@@ -21,6 +20,7 @@
 
 #include "knowhere/index/IndexType.h"
 #include "knowhere/index/vector_index/adapter/VectorAdapter.h"
+#include "unittest/benchmark/benchmark_base.h"
 
 /*****************************************************
  * To run this test, please download the HDF5 from
@@ -55,129 +55,11 @@ static const char* METRIC_TAN_STR = "tanimoto";
  * SIFT     128         1,000,000   10,000      100         Euclidean   HDF5 (501MB)
  *************************************************************************************/
 
-#define CALC_TIME_SPAN(X)       \
-    double t_start = elapsed(); \
-    X;                          \
-    double t_diff = elapsed() - t_start;
-
-class Benchmark_sift : public ::testing::Test {
+class Benchmark_hdf5 : public Benchmark_base {
  public:
-    void
-    normalize(float* arr, int32_t nq, int32_t dim) {
-        for (int32_t i = 0; i < nq; i++) {
-            double vecLen = 0.0, inv_vecLen = 0.0;
-            for (int32_t j = 0; j < dim; j++) {
-                double val = arr[i * dim + j];
-                vecLen += val * val;
-            }
-            inv_vecLen = 1.0 / std::sqrt(vecLen);
-            for (int32_t j = 0; j < dim; j++) {
-                arr[i * dim + j] = (float)(arr[i * dim + j] * inv_vecLen);
-            }
-        }
-    }
-
-    double
-    elapsed() {
-        struct timeval tv;
-        gettimeofday(&tv, nullptr);
-        return tv.tv_sec + tv.tv_usec * 1e-6;
-    }
-
-    double
-    get_time_diff() {
-        return elapsed() - T0_;
-    }
-
     void
     set_ann_test_name(const char* test_name) {
         ann_test_name_ = test_name;
-    }
-
-    /* Distance checking is only meaningful for binary index or FLAT index */
-    void
-    CheckDistance(const knowhere::MetricType& metric_type,
-                  const int64_t* ids,
-                  const float* distances,
-                  const size_t* lims,
-                  int32_t nq) {
-        const float FLOAT_DIFF = 0.00001;
-        const bool is_L2 = (metric_type == knowhere::metric::L2);
-        for (int32_t i = 0; i < nq; i++) {
-            std::unordered_set<int32_t> gt_ids_set(gt_ids_ + gt_lims_[i], gt_ids_ + gt_lims_[i + 1]);
-            std::unordered_map<int32_t, float> gt_map;
-            for (auto j = gt_lims_[i]; j < gt_lims_[i + 1]; j++) {
-                gt_map[gt_ids_[j]] = gt_dist_[j];
-            }
-            for (auto j = lims[i]; j < lims[i + 1]; j++) {
-                if (gt_ids_set.count(ids[j]) > 0) {
-                    float dist = (is_L2) ? std::sqrt(distances[j]) : distances[j];
-                    ASSERT_LT(std::abs(dist - gt_map[ids[j]]), FLOAT_DIFF);
-                }
-            }
-        }
-    }
-
-    float
-    CalcRecall(const int64_t* ids, int32_t nq, int32_t k) {
-        int32_t min_k = std::min(gt_k_, k);
-        int32_t hit = 0;
-        for (int32_t i = 0; i < nq; i++) {
-            std::unordered_set<int32_t> ground(gt_ids_ + i * gt_k_, gt_ids_ + i * gt_k_ + min_k);
-            for (int32_t j = 0; j < min_k; j++) {
-                auto id = ids[i * k + j];
-                if (ground.count(id) > 0) {
-                    hit++;
-                }
-            }
-        }
-        return (hit * 1.0f / (nq * min_k));
-    }
-
-    float
-    CalcRecall(const int64_t* ids, int32_t nq_start, int32_t step, int32_t k) {
-        assert(nq_start + step <= 10000);
-        int32_t min_k = std::min(gt_k_, k);
-        int32_t hit = 0;
-        for (int32_t i = 0; i < step; i++) {
-            std::unordered_set<int32_t> ground(gt_ids_ + (i + nq_start) * gt_k_,
-                                               gt_ids_ + (i + nq_start) * gt_k_ + min_k);
-            for (int32_t j = 0; j < min_k; j++) {
-                auto id = ids[i * k + j];
-                if (ground.count(id) > 0) {
-                    hit++;
-                }
-            }
-        }
-        return (hit * 1.0f / (step * min_k));
-    }
-
-    float
-    CalcRecall(const int64_t* ids, const size_t* lims, int32_t nq) {
-        int32_t hit = 0;
-        for (int32_t i = 0; i < nq; i++) {
-            std::unordered_set<int32_t> gt_ids_set(gt_ids_ + gt_lims_[i], gt_ids_ + gt_lims_[i + 1]);
-            for (auto j = lims[i]; j < lims[i + 1]; j++) {
-                if (gt_ids_set.count(ids[j]) > 0) {
-                    hit++;
-                }
-            }
-        }
-        return (hit * 1.0f / gt_lims_[nq]);
-    }
-
-    float
-    CalcAccuracy(const int64_t* ids, const size_t* lims, int32_t nq) {
-        int32_t hit = 0;
-        for (int32_t i = 0; i < nq; i++) {
-            std::unordered_set<int64_t> ids_set(ids + lims[i], ids + lims[i + 1]);
-            for (auto j = gt_lims_[i]; j < gt_lims_[i + 1]; j++) {
-                if (ids_set.count(gt_ids_[j]) > 0) {
-                    hit++;
-                }
-            }
-        }
-        return (hit * 1.0f / lims[nq]);
     }
 
     void
@@ -315,28 +197,6 @@ class Benchmark_sift : public ::testing::Test {
 
         gt_dist_ = (float*)hdf5_read(ann_file_name, HDF5_DATASET_DISTANCES, H5T_FLOAT, cols, rows);
         assert((cols == gt_lims_[nq_] && rows == 1) || !"incorrect dims of ground truth distances");
-    }
-
-    void
-    free_all() {
-        if (xb_ != nullptr) {
-            delete[](float*) xb_;
-        }
-        if (xq_ != nullptr) {
-            delete[](float*) xq_;
-        }
-        if (gt_radius_ != nullptr) {
-            delete[] gt_radius_;
-        }
-        if (gt_lims_ != nullptr) {
-            delete[] gt_lims_;
-        }
-        if (gt_ids_ != nullptr) {
-            delete[] gt_ids_;
-        }
-        if (gt_dist_ != nullptr) {
-            delete[] gt_dist_;
-        }
     }
 
  protected:
@@ -514,17 +374,6 @@ class Benchmark_sift : public ::testing::Test {
     }
 
  protected:
-    double T0_;
     std::string ann_test_name_ = "";
     std::string metric_str_;
-    int32_t dim_;
-    void* xb_ = nullptr;
-    void* xq_ = nullptr;
-    int32_t nb_;
-    int32_t nq_;
-    float* gt_radius_ = nullptr; // ground-truth radius
-    int32_t* gt_lims_ = nullptr; // ground-truth lims
-    int32_t* gt_ids_ = nullptr;  // ground-truth labels
-    float* gt_dist_ = nullptr;   // ground-truth distances
-    int32_t gt_k_;
 };
