+++ unittest/benchmark/hdf5/benchmark_knowhere_float_range.cpp
@@ -12,10 +12,10 @@
 #include <gtest/gtest.h>
 #include <vector>
 
-#include "unittest/benchmark/benchmark_knowhere.h"
+#include "benchmark_knowhere.h"
 #include "unittest/range_utils.h"
 
-class Benchmark_knowhere_float_range : public Benchmark_knowhere {
+class Benchmark_knowhere_float_range : public Benchmark_knowhere, public ::testing::Test {
  public:
     void
     test_idmap(const knowhere::Config& cfg) {
